Updating the wireing of the LAYOUT macro.
It turns out that the LAYOUT macro which used to be called KEYMAP
changed around some of the terminonlogy about how the columns were
numbered.  This in turn changed how the existing key mapping worked with
the new firmware.  I have altered the layout to correctly model my
column wiring and moved around the key mappings back to what I was
expecing.